docs: update menu
