+++ docs/demo/menu/section-fat.jsx
@@ -2,7 +2,7 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import Menu from '../../../components/menu'
 const prefix = 'menu-fat'
-const desc = '二级菜单项，以分组的形式展示，分组数不宜'
+const desc = '二级菜单项，以分组的形式展示，分组数不宜过多'
 const code = `import React from 'react'
 import Menu from '@hi-ui/hiui/es/menu'\n
 class Demo extends React.Component {
@@ -125,11 +125,6 @@ class Demo extends React.Component {
 }`
 
 const DemoFat = () => (
-  <DocViewer
-    code={code}
-    desc={desc}
-    scope={{ Menu }}
-    prefix={prefix}
-  />
+  <DocViewer code={code} desc={desc} scope={{ Menu }} prefix={prefix} />
 )
 export default DemoFat
