+++ docs/demo/menu/section-vertical.jsx
@@ -2,19 +2,24 @@ import React from 'react'
 import DocViewer from '../../../libs/doc-viewer'
 import Menu from '../../../components/menu'
 const prefix = 'menu-vertical'
+const desc = '当导航的菜单项和层级较多时适用，可收起'
+const rightOptions = ['基础', '可同时展开多个', '可收起']
 
-const code = `import React from 'react'
+const code = [
+  {
+    code: `import React from 'react'
 import Menu from '@hi-ui/hiui/es/menu'\n
 class Demo extends React.Component {
   render(){
-    const datas = [
+    const data = [
       {
         content: '电视',
-        icon: 'internet',
+        icon: 'alarm',
         id: 1
       },
       {
         content: '小米MIX',
+        icon: 'internet',
         id: 2
       },
       {
@@ -56,7 +61,6 @@ class Demo extends React.Component {
           },
           {
             content: '红米',
-            icon: 'phone',
             id: 'hongmi'
           },
           {
@@ -88,6 +92,7 @@ class Demo extends React.Component {
       },
       {
         content: '超长超长超长字符超长超长超长字符',
+        icon: 'document',
         id: 4
       }
     ]
@@ -98,17 +103,234 @@ class Demo extends React.Component {
           activeId={'xiaomi9'}
           onClick={(id, prevId)=>console.log('-----click', id, prevId)}
           onClickSubMenu={index => console.log('-----onClickSubMenu', index)}
-          data={datas}
+          data={data}
         />
       </div>
     )
   }
-}`
+}`,
+    opt: ['基础']
+  },
+  {
+    code: `import React from 'react'
+import Menu from '@hi-ui/hiui/es/menu'\n
+class Demo extends React.Component {
+  render(){
+    const data = [
+      {
+        content: '电视',
+        icon: 'alarm',
+        id: 1
+      },
+      {
+        content: '小米MIX',
+        icon: 'internet',
+        id: 2
+      },
+      {
+        content: '手机',
+        icon: 'phone',
+        children: [
+          {
+            content: '小米',
+            children: [
+              {
+                content: '小米9',
+                id: 'xiaomi9'
+              },
+              {
+                content: '小米8',
+                id: 'xiaomi8'
+              },
+              {
+                content: '小米7',
+                id: 'xiaomi7'
+              },
+              {
+                content: '小米6',
+                id: 'xiaomi6'
+              },
+              {
+                content: '小米5',
+                id: 'xiaomi5'
+              },
+              {
+                content: '小米4',
+                id: 'xiaomi4'
+              },
+              {
+                content: '小米3',
+                id: 'xiaomi3'
+              }
+            ]
+          },
+          {
+            content: '红米',
+            id: 'hongmi'
+          },
+          {
+            content: '小米note',
+            children: [
+              {
+                content: '小米 note7',
+                id: 'xiaomi note7'
+              },
+              {
+                content: '小米 note6',
+                id: 'xiaomi note6'
+              },
+              {
+                content: '小米 note5',
+                id: 'xiaomi note5'
+              },
+              {
+                content: '小米 note4',
+                id: 'xiaomi note4'
+              },
+              {
+                content: '小米 note3',
+                id: 'xiaomi note3'
+              }
+            ]
+          }
+        ]
+      },
+      {
+        content: '超长超长超长字符超长超长超长字符',
+        icon: 'document',
+        id: 4
+      }
+    ]
+    return(
+      <div>
+        <Menu
+          accordion={false}
+          placement="vertical"
+          activeId={'xiaomi9'}
+          onClick={(id, prevId)=>console.log('-----click', id, prevId)}
+          onClickSubMenu={index => console.log('-----onClickSubMenu', index)}
+          data={data}
+        />
+      </div>
+    )
+  }
+}`,
+    opt: ['可同时展开多个']
+  },
+  {
+    code: `import React from 'react'
+import Menu from '@hi-ui/hiui/es/menu'\n
+class Demo extends React.Component {
+  render(){
+    const data = [
+      {
+        content: '电视',
+        icon: 'alarm',
+        id: 1
+      },
+      {
+        content: '小米MIX',
+        icon: 'internet',
+        id: 2
+      },
+      {
+        content: '手机',
+        icon: 'phone',
+        children: [
+          {
+            content: '小米',
+            children: [
+              {
+                content: '小米9',
+                id: 'xiaomi9'
+              },
+              {
+                content: '小米8',
+                id: 'xiaomi8'
+              },
+              {
+                content: '小米7',
+                id: 'xiaomi7'
+              },
+              {
+                content: '小米6',
+                id: 'xiaomi6'
+              },
+              {
+                content: '小米5',
+                id: 'xiaomi5'
+              },
+              {
+                content: '小米4',
+                id: 'xiaomi4'
+              },
+              {
+                content: '小米3',
+                id: 'xiaomi3'
+              }
+            ]
+          },
+          {
+            content: '红米',
+            id: 'hongmi'
+          },
+          {
+            content: '小米note',
+            children: [
+              {
+                content: '小米 note7',
+                id: 'xiaomi note7'
+              },
+              {
+                content: '小米 note6',
+                id: 'xiaomi note6'
+              },
+              {
+                content: '小米 note5',
+                id: 'xiaomi note5'
+              },
+              {
+                content: '小米 note4',
+                id: 'xiaomi note4'
+              },
+              {
+                content: '小米 note3',
+                id: 'xiaomi note3'
+              }
+            ]
+          }
+        ]
+      },
+      {
+        content: '超长超长超长字符超长超长超长字符',
+        icon: 'document',
+        id: 4
+      }
+    ]
+    return(
+      <div>
+        <Menu
+          showCollapse
+          placement="vertical"
+          activeId={'xiaomi9'}
+          onClick={(id, prevId)=>console.log('-----click', id, prevId)}
+          onClickSubMenu={index => console.log('-----onClickSubMenu', index)}
+          data={data}
+        />
+      </div>
+    )
+  }
+}`,
+    opt: ['可收起']
+  }
+]
 const DemoVertical = () => (
   <DocViewer
     code={code}
     scope={{ Menu }}
     prefix={prefix}
+    desc={desc}
+    rightOptions={rightOptions}
   />
 )
 export default DemoVertical
