+++ docs/zh-CN/components/menu.mdx
@@ -4,8 +4,9 @@
 
 ## 何时使用
 
-- 顶部导航
-- 侧边栏导航
+顶部导航
+
+侧边栏导航
 
 ## 基础用法
 
@@ -13,46 +14,34 @@ import DemoRow from '../../demo/menu/section-row.jsx'
 
 <DemoRow />
 
-## 分组菜单
+## 分组用法
 
 import DemoFat from '../../demo/menu/section-fat.jsx'
 
 <DemoFat />
 
-## 垂直菜单
+## 垂直方向
 
 import DemoVertical from '../../demo/menu/section-vertical.jsx'
 
 <DemoVertical />
 
-## 垂直菜单可同时展开多个
-
-import DemoNoAccordion from '../../demo/menu/section-no-accordion.jsx'
-
-<DemoNoAccordion />
-
-## 垂直菜单可收起
-
-import DemoPack from '../../demo/menu/section-pack.jsx'
-
-<DemoPack />
-
 import { Badge } from '@libs'
 
 ## Props
 
-| 参数            | 说明                                 | 类型                             | 可选值                     | 默认值     |
-| --------------- | ------------------------------------ | -------------------------------- | -------------------------- | ---------- |
-| data            | 菜单项数据源                         | DataItem[]                       | -                          | -          |
-| activeId        | 激活的菜单项 id                      | string                           | -                          | -          |
-| placement       | 设置菜单水平或垂直展示               | string                           | 'horizontal' \| 'vertical' | 'vertical' |
-| collapsed       | 是否收起子菜单，菜单垂直展示时有效   | boolean                          | true \| false              | false      |
-| showCollapse    | 是否显示收缩开关，菜单垂直展示时有效 | boolean                          | true \| false              | true       |
-| showAllSubMenus | 是否以胖菜单的形式展开所有子菜单（仅在水平菜单时有效）                     | boolean                          | true \| false              | false      |
-| accordion       | 手风琴模式，菜单水平展示时有效       | boolean                          | true \| false              | true       |
-| onClick         | 点击菜单选项时的回调                 | (activeId, prevActiveId) => void | -                          | -          |
-| onClickSubMenu  | 点击父菜单项时的回调                 | (subMenuIndexs) => void          | -                          | -          |
-| onCollapse      | 点击收缩开关时的回调                 | (collapsed: boolean) => void     | -                          | -          |
+| 参数            | 说明                                                   | 类型                             | 可选值                     | 默认值     |
+| --------------- | ------------------------------------------------------ | -------------------------------- | -------------------------- | ---------- |
+| data            | 菜单项数据源                                           | DataItem[]                       | -                          | -          |
+| activeId        | 激活的菜单项 id                                        | string                           | -                          | -          |
+| placement       | 设置菜单水平或垂直展示                                 | string                           | 'horizontal' \| 'vertical' | 'vertical' |
+| collapsed       | 是否收起子菜单，菜单垂直展示时有效                     | boolean                          | true \| false              | false      |
+| showCollapse    | 是否显示收缩开关，菜单垂直展示时有效                   | boolean                          | true \| false              | true       |
+| showAllSubMenus | 是否以胖菜单的形式展开所有子菜单（仅在水平菜单时有效） | boolean                          | true \| false              | false      |
+| accordion       | 手风琴模式，菜单水平展示时有效                         | boolean                          | true \| false              | true       |
+| onClick         | 点击菜单选项时的回调                                   | (activeId, prevActiveId) => void | -                          | -          |
+| onClickSubMenu  | 点击父菜单项时的回调                                   | (subMenuIndexs) => void          | -                          | -          |
+| onCollapse      | 点击收缩开关时的回调                                   | (collapsed: boolean) => void     | -                          | -          |
 
 ## Type
